@@ -7,28 +7,36 @@
   },
   "dependencies": {
     "@babel/core": "^7.14.6",
-    "@emotion/react": "^11.10.5",
+    "@emotion/react": "^11.10.6",
     "@emotion/server": "^11.10.0",
-    "@mantine/core": "^5.9.2",
-    "@mantine/hooks": "^5.9.2",
-    "@mantine/next": "^5.9.2",
+    "@mantine/core": "^6.0.1",
+    "@mantine/hooks": "^6.0.1",
+    "@mantine/next": "^6.0.1",
     "axios": "^1.2.2",
     "babel-preset-next": "^1.4.0",
     "dayjs": "^1.11.7",
-    "discord-oauth2": "^2.7.0",
     "formik": "^2.2.9",
     "mongodb": "^4.13.0",
     "mongoose": "^6.8.3",
-    "next": "^13.1.1",
-    "next-auth": "^4.18.8",
+    "next": "^13.2.4",
+    "next-auth": "^4.20.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "styled-components": "^5.3.0",
     "yup": "^0.32.9"
   },
   "license": "MIT",
+  "resolutions": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
+  },
   "devDependencies": {
     "@types/react": "^18.0.26",
+    "@typescript-eslint/eslint-plugin": "^5.48.1",
+    "@typescript-eslint/parser": "^5.48.1",
+    "eslint": "^8.31.0",
+    "eslint-plugin-react": "^7.31.11",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "typescript": "^4.3.5"
   }
 }
