@@ -1,31 +1,24 @@
 import React, { useState } from "react";
 import * as Yup from "yup";
 import { withFormik } from "formik";
-import { ranks, regions, roles } from "../util/datalist";
+import { fullRanks, mentorFormQuestions, regions } from "../util/datalist";
 import { FormSelect, FormTextField, StyledForm } from "./Styles";
 import axios from "axios";
 import { useRouter } from "next/router";
-import { Textarea } from "@mantine/core";
+import { Button, Textarea } from "@mantine/core";
 
 const FormEnhancer = withFormik({
   validationSchema: Yup.object().shape({
     rank: Yup.string().required("Please select your Rank "),
     summonerName: Yup.string().required("Please enter your Summoner Name"),
-    role: Yup.string().required("Please select your role"),
     region: Yup.string().required("Please select your region"),
-    timezone: Yup.string().required("Please select your timezone"),
+    appReason: Yup.string().required("*Required"),
+    loseMatchupEx: Yup.string().required("*Required"),
+    rebuttalEx: Yup.string().required("*Required"),
+    winConEx: Yup.string().required("*Required"),
   }),
-  handleSubmit: () => {},
   validateOnMount: true,
 });
-
-const timeZone = () => {
-  const zones = [];
-  for (let i = -11; i <= 12; i++) {
-    zones.push(`UTC ${i}`);
-  }
-  return zones;
-};
 
 const MentorRequestForm = (props) => {
   const { isValid, values, errors, handleChange, handleBlur, setFieldValue } =
@@ -33,32 +26,38 @@ const MentorRequestForm = (props) => {
 
   const router = useRouter();
   const [loading, setLoading] = useState(false);
-  const handleSubmit = async () => {
+
+  const handleSubmit = () => {
     setLoading(true);
     if (isValid) {
-      await axios
-        .post("/api/request/create", values)
+      axios
+        .post("/api/user/application", values)
         .then(() => {
-          alert("Request has been sent!");
+          alert("Request has been sent! We will reach you back in 1-2 weeks");
           router.push("/");
         })
         .catch((error) => {
-          alert(
-            "Error. Please check your form. If this issue persists, please contact the Mod team"
-          );
+          alert(error.response.data.error);
+          setLoading(false);
         });
     }
-    setLoading(false);
   };
 
   const onChange = (obj, field) => {
-    setFieldValue(field, obj.value);
+    setFieldValue(field, obj);
   };
 
-  const zones = timeZone();
-
   return (
-    <StyledForm onSubmit={handleSubmit}>
+    <StyledForm>
+      <FormTextField
+        title="Summoner name"
+        id="summonerName"
+        placeholder="Only put your main account"
+        value={values.summonerName}
+        onChange={handleChange}
+        onBlur={handleBlur}
+        error={errors.summonerName}
+      />
       <FormSelect
         title="Region"
         name="region"
@@ -66,57 +65,26 @@ const MentorRequestForm = (props) => {
         error={errors.region}
         onChange={onChange}
       />
-      <FormTextField
-        title="Summoner name"
-        id="summonerName"
-        placeholder="Summoner name"
-        value={values.summonerName}
-        onChange={handleChange}
-        onBlur={handleBlur}
-        errorText={errors.summonerName}
-      />
       <FormSelect
-        title="Role"
-        name="role"
-        options={roles}
+        title="Rank"
+        name="rank"
+        options={fullRanks()}
         onChange={onChange}
-        error={errors.role}
+        error={errors.rank}
       />
-      <MultiSelect
-        label={"Champions"}
-        data={championList}
-        error={errors.champions}
-        searchable
-        onChange={(e) => onChange(e, "champions")}
-      />
-      <Textarea
-        label="What made you want to apply as a mentor?"
-        id="appReason"
-        value={values.appReason}
-        error={errors.appReason}
-        onBlur={handleBlur}
-      />
+      {mentorFormQuestions.map((question, i) => (
+        <Textarea
+          key={`MentorFormQuestion${i}`}
+          onChange={handleChange}
+          label={question.title}
+          id={question.field}
+          error={errors[question.field]}
+        />
+      ))}
 
-      <FormSelect
-        title="Timezone"
-        name="timezone"
-        options={zones}
-        onChange={onChange}
-        error={errors.timezone}
-      />
-
-      <FormTextField
-        title="Any additional information you would like the mentors to know (if nothing, leave blank)"
-        id="info"
-        placeholder="Info"
-        value={values.info}
-        onChange={handleChange}
-        onBlur={handleBlur}
-      />
-
-      <button onClick={handleSubmit} disabled={loading}>
-        Send Request
-      </button>
+      <Button onClick={handleSubmit} disabled={loading}>
+        Send application
+      </Button>
     </StyledForm>
   );
 };
