@@ -1,4 +1,5 @@
 import { Button } from "@mantine/core";
+import React from "react";
 
 export default function Terms({ setTerms }) {
   return (
@@ -43,7 +44,7 @@ export default function Terms({ setTerms }) {
         least 25 total ranked games played this season - Must be a fairly close
         game - no one-sided stomps - Must be serious about improvement
       </p>
-      <Button variant="outlined" onClick={() => setTerms(true)}>
+      <Button onClick={() => setTerms(true)}>
         By clicking this box, you acknowledge that you have read all of the
         above.
       </Button>
