@@ -1,13 +1,18 @@
 import Header from "../components/header";
 import Footer from "../components/footer";
-import { PageWrappaer } from "./Styles";
+import React from "react";
+import { PageWrapper } from "./Styles";
+import Head from "next/head";
 
 export default function Layout({ children }) {
   return (
-    <PageWrappaer>
+    <PageWrapper>
+      <Head>
+        <title>Summoner School Mentor Requests</title>
+      </Head>
       <Header />
       <main>{children}</main>
       <Footer />
-    </PageWrappaer>
+    </PageWrapper>
   );
 }
