@@ -1,62 +1,50 @@
-import axios from "axios";
-import { useSession } from "next-auth/react";
-import { useState, useEffect } from "react";
+import React from "react";
 import Layout from "../components/layout";
 import Link from "next/link";
-import { useRouter } from "next/dist/client/router";
 import MentorForm from "../components/MentorForm";
-import { Button } from "@mantine/core";
 import { FormWrapper } from "../components/Styles";
+import { getToken } from "next-auth/jwt";
+import { checkPendingApp } from "../util/databaseAccess";
 
-export default function Mentors() {
-  const [content, setContent] = useState("");
-  const { data: session, status } = useSession();
-  const loading = status === "loading";
-  const router = useRouter();
-
-  const handleOnClick = async () => {
-    const data = (await axios.post("/api/user/mentor", session)).data;
-    switch (data) {
-      case 0:
-        setContent(
-          "request has been sent. Be sure to let the Admins know on the server"
-        );
-        break;
-      case 1:
-        setContent(
-          <>
-            Mentor already registered! You can head over to{" "}
-            <Link href="https://ssmentor-request.vercel.app/mentors">
-              this link
-            </Link>{" "}
-            to now see the requests
-          </>
-        );
-        break;
-      case 2:
-        setContent(
-          "request has already been sent. We'll get to it when we get to it"
-        );
-        break;
-    }
-  };
-
-  useEffect(async () => {
-    if (!loading)
-      if (!session) {
-        router.push("/api/auth/signin");
-      }
-  }, [loading]);
-
-  return (
-    <Layout>
-      <FormWrapper>
-        {/* <MentorForm /> */}
-        <Button onClick={handleOnClick}>Click here to apply for mentor</Button>
-        <br />
-        <MentorForm />
-        {content}
-      </FormWrapper>
-    </Layout>
+export default function Apply({ isRegistered, isMentor }) {
+  let content = (
+    <FormWrapper>
+      <MentorForm />
+    </FormWrapper>
   );
+
+  if (isRegistered)
+    content =
+      "request has already been sent. We'll get to it when we get to it";
+
+  if (isMentor)
+    content = (
+      <>
+        Mentor already registered! You can head over to{" "}
+        <Link href="https://ssmentor-request.vercel.app/mentors">
+          this link
+        </Link>{" "}
+        to now see the requests
+      </>
+    );
+
+  return <Layout>{content}</Layout>;
+}
+
+export async function getServerSideProps({ req }) {
+  const token = await getToken({ req });
+
+  if (!token) {
+    return {
+      redirect: {
+        destination: "/api/auth/signin",
+        permanent: false,
+      },
+    };
+  }
+  const isRegistered = await checkPendingApp(token.sub);
+
+  return {
+    props: { isRegistered },
+  };
 }
