@@ -1,37 +1,15 @@
 import { connectToDatabase } from "./mongodb";
 import Request from "../models/requestModel";
+import MentorApp from "../models/mentorAppModel";
 import User from "../models/userModel";
-import { isMentor } from "./helper";
-import mongoose from "mongoose";
-import { ObjectID } from "mongodb";
+import { checkAdmin, checkMentor, checkReviewer } from "./helper";
+import { ObjectId } from "mongodb";
 
 connectToDatabase();
 
 export async function testRequest() {
-  const requests = await Request.find({
-    archived: { $ne: true },
-    $or: [
-      {
-        $and: [
-          { status: "Completed" },
-          {
-            completed: { $gte: threeMonthsAgo },
-          },
-        ],
-      },
-      {
-        $and: [
-          { status: "Problem" },
-          {
-            createdAt: { $gte: threeMonthsAgo },
-          },
-        ],
-      },
-      { status: { $nin: ["Completed", "Problem"] } },
-    ],
-  });
-
-  return requests;
+  await User.updateMany({ userType: "mentor" }, { isMentor: true });
+  return "Bonga";
 }
 
 export async function getAllRequests() {
@@ -65,48 +43,94 @@ export async function getAllRequests() {
       { status: { $nin: ["Completed", "Problem"] } },
     ],
   })
-    .populate({ path: "mentor", model: "User" })
-    // .limit(15)
+    .populate({ path: "mentor", model: "User", select: "discordName" })
+    // .limit(5)
     .sort({ createdAt: 1 })
-    .then((items) => cleaner(items));
+    .lean();
+  // .then((items) => cleaner(items));
+  return requests;
+}
+
+export async function getTypeRequests(type) {
+  const requests = await Request.find({ status: type })
+    .populate({ path: "mentor", model: "User", select: "discordName" })
+    .sort({ createdAt: 1 })
+    .lean();
+  // .then((items) => cleaner(items));
   return requests;
 }
 
 export async function getAllMentors() {
-  return await User.find({ userType: { $nin: ["god", "admin"] } }).then(
-    (mentors) => userCleaner(mentors)
+  const mentors = await User.find({
+    $or: [{ userType: "mentor" }, { isMentor: true }],
+  }).lean();
+
+  return cleaner(mentors);
+}
+
+export async function getMentorDetails(id) {
+  return await User.findOne({ discordId: id }).then((mentor) =>
+    singleCleaner(mentor)
   );
 }
 
 export async function getAllUsers() {
-  return await User.find().then((mentors) => userCleaner(mentors));
+  const users = await User.find();
+  const newUsers = await Promise.all(
+    users.map(async (mentor) => {
+      const lastTaken = (
+        await Request.findOne({ mentor: mentor._id }).sort({
+          accepted: -1,
+        })
+      )?.accepted;
+
+      const lastCompleted = (
+        await Request.findOne({
+          mentor: mentor._id,
+          status: "Completed",
+        }).sort({ completed: -1 })
+      )?.completed;
+
+      return {
+        ...mentor.toObject(),
+        lastTaken: lastTaken || null,
+        lastCompleted: lastCompleted || null,
+      };
+    })
+  );
+  return cleaner(newUsers);
 }
 
-export async function getMentorRequests(discordId) {
-  const mentor = await User.find({ discordId });
+export async function getMenteeRequestsByDiscordId(id) {
+  return await Request.find({ discordId: id }).then((requests) =>
+    cleaner(requests)
+  );
+}
+
+export async function getMentorRequests(_id) {
   const requests = await Request.find({
-    mentor: mongoose.Types.ObjectId(userId),
+    mentor: ObjectId(_id),
   })
     .sort({ completed: -1 })
+    .lean()
     .then((items) => cleaner(items));
 
-  return { mentor, requests };
+  return requests;
 }
 
-export async function isRequestPending(session) {
+export async function isRequestPending(id) {
   const temp = await Request.find({
-    $query: { discordId: session.user.id },
+    discordId: id,
   })
     .sort({ createdAt: -1 })
     .limit(1);
 
   const request = temp[0];
 
-  if (request) {
-    if (!request.completed || wasMonthAgo(request.completed)) {
-      return true;
-    }
+  if (request && (!request.completed || wasMonthAgo(request.completed))) {
+    return true;
   }
+
   return false;
 }
 
@@ -117,71 +141,94 @@ function wasMonthAgo(completed) {
   return monthAgo < then;
 }
 
-export async function getSessionUser(session) {
-  return await User.findOne({ discordId: session.user.id });
+export async function getUserById(id) {
+  return await User.findOne({ discordId: id })
+    .select("-__v -createdAt -updatedAt")
+    .lean()
+    .then((user) => singleCleaner(user));
 }
 
-export async function tryRegisterMentor(session) {
-  const user = await getSessionUser(session);
-  if (!user) {
-    const newUser = new User({
-      discordName: `${session.user.name}#${session.user.discriminator}`,
-      discordId: session.user.id,
-    });
-    await newUser.save();
-    return 0;
+export async function isUserAdmin(id) {
+  const user = await getUserById(id);
+  return checkAdmin(user);
+}
+
+export async function isUserMentor(id) {
+  return checkMentor(await getUserById(id));
+}
+
+export async function isUserReviewer(id) {
+  return checkReviewer(await getUserById(id));
+}
+
+export async function tryRegisterMentor(user) {
+  const registeredUser = await getUserById(user.discordId);
+  if (registeredUser) {
+    throw "mentor is already registered";
   }
-  if (isMentor(user)) return 1;
-  return 2;
+
+  const newUser = new User({
+    ...user,
+    isMentor: true,
+  });
+  await newUser.save();
+  return "mentor created!";
 }
 
 export async function editUser(body) {
   try {
-    const { userID, command } = body;
-    const user = await User.findById(userID);
+    const { userId, command } = body;
+    const user = await User.findById(userId);
     switch (command) {
       case "SET_MENTOR":
-        user.userType = "mentor";
-        break;
-      case "SET_USER":
-        user.userType = "user";
+        user.isMentor = body.bool;
         break;
       case "SET_ADMIN":
-        user.userType = "admin";
+        user.isAdmin = body.bool;
         break;
+      case "SET_REVIEWER":
+        user.isReviewer = body.bool;
+        break;
+      case "SET_PEAKRANK":
+        user.peakRank = body.rank;
+        break;
+      case "SET_CHAMPIONS":
+        user.bestChampions = body.value;
+        break;
+      case "SET_REGION":
+        user.mentorRegion = body.value;
+        break;
+      case "SET_ROLES":
+        user.preferredRoles = body.value;
     }
     await user.save();
     return `Sucessfully edited to ${command}`;
   } catch {
-    return `Error trying to run ${command}`;
+    return `Error trying to run ${body.command}`;
   }
 }
 
-export async function changeRequest(body) {
-  const { user } = body.session;
-  const now = new Date();
-
-  const request = await Request.findOne({ _id: body.id });
-  const mentor = await User.findOne({
-    discordId: user.id,
-    userType: { $ne: "user" },
-  });
-
-  if (!request || !mentor) {
-    throw { error: "I uh... what?" };
+export async function deleteUser({ user }) {
+  try {
+    const response = await User.deleteOne({ _id: ObjectId(user) });
+    console.log(response);
+    return "User deletion success!";
+  } catch {
+    return "Error trying to delete user";
   }
+}
 
-  if (body.type == "delete") {
-    await Request.deleteOne({
-      _id: ObjectID(request._id),
-    });
+export async function changeRequest({ body, user }) {
+  const now = new Date();
+  const request = await Request.findById(body.id);
 
-    return "Success";
+  if (!request) {
+    throw { error: "I uh... what?" };
   }
 
   switch (body.type) {
     case "status":
-      request.mentor = mentor._id;
+      request.mentor = user._id;
       request.status = body.value;
 
       switch (body.value) {
@@ -189,7 +236,6 @@ export async function changeRequest(body) {
           request.accepted = now;
           break;
         case "Not Accepted":
-          request.mentor = null;
           request.accepted = null;
           request.completed = null;
           break;
@@ -214,55 +260,112 @@ export async function changeRequest(body) {
   await request.save();
 }
 
-export async function createRequest(body) {
-  const {
-    session: { user },
-    values,
-  } = body;
+export async function deleteRequest(id) {
+  const state = await Request.findByIdAndDelete(id);
+  console.log(state);
+}
 
+export async function createRequest({ values, user }) {
   const hasRequest = await Request.findOne({
-    discordId: values.discordId,
+    discordId: user.sub,
+    status: { $ne: "Completed" },
   });
 
   if (hasRequest) {
-    throw res.status(401).send("user already has request");
+    throw { error: "user already has request" };
   }
 
   const request = new Request({
     ...values,
     discordName: `${user.name}#${user.discriminator}`,
-    discordId: user.id,
+    discordId: user.sub,
   });
-
   return await request.save();
 }
 
-function userCleaner(mentors) {
-  return mentors.map((mentor) => ({
-    discordId: mentor.discordId,
-    discordName: mentor.discordName,
-    _id: mentor._id.toString(),
-    userType: mentor.userType,
-  }));
+export async function checkPendingApp(id) {
+  return (await MentorApp.findOne({ discordId: id })) && true;
+}
+
+export async function createApp(user, details) {
+  const app = new MentorApp({
+    ...details,
+    discordId: user.sub,
+    discordName: `${user.name}#${user.discriminator}`,
+  });
+
+  return await app.save();
+}
+
+export async function getAllApps() {
+  return await MentorApp.find()
+    .select("-createdAt -updatedAt -__v")
+    .lean()
+    .then((data) => cleaner(data));
+}
+
+export async function voteOnApp({ id, vote, reviewer }) {
+  const app = await MentorApp.findOne({ discordId: id });
+
+  app.yay = app.yay.filter((id) => id != reviewer);
+  app.nay = app.nay.filter((id) => id != reviewer);
+  app.meh = app.meh.filter((id) => id != reviewer);
+  if (app.yay > 3) tryRegisterMentor();
+  app[vote].push(reviewer);
+  await app.save();
+}
+
+export async function deleteApp(id) {
+  const response = await MentorApp.findOneAndDelete({ discordId: id });
+  console.log(response);
+}
+
+export async function processApp(id) {
+  const resp = await MentorApp.findOneAndUpdate(
+    { discordId: id },
+    { processed: true }
+  );
+  console.log(resp);
+}
+
+function singleCleaner(mentor) {
+  for (let key in mentor) {
+    const varMentor = mentor[key];
+    if (
+      varMentor &&
+      typeof varMentor == "object" &&
+      !varMentor.length &&
+      varMentor.length != 0
+    ) {
+      mentor[key] = varMentor.toString();
+    } else if (key == "mentor") mentor[key] = varMentor?.discordName;
+    else if (varMentor === undefined) mentor[key] = null;
+  }
+
+  return mentor;
 }
 
 function cleaner(items) {
-  return items.map((item) => ({
-    id: item._id.toString(),
-    status: item.status,
-    rank: item.rank,
-    region: item.region,
-    summonerName: item.summonerName || item.opgg || "what?",
-    role: item.role,
-    champions: item.champions || null,
-    timezone: item.timezone,
-    info: item.info || null,
-    createdAt: item.createdAt.toISOString(),
-    discordName: item.discordName,
-    discordId: item.discordId,
-    accepted: item.accepted?.toISOString() || null,
-    completed: item.completed?.toISOString() || null,
-    mentor: item.mentor?.discordName || null,
-    remarks: item.remarks || null,
-  }));
+  return items.map((item) => {
+    return singleCleaner(item);
+  });
+
+  // return items.map((item) => ({
+  //   id: item._id.toString(),
+  //   status: item.status,
+  //   rank: item.rank,
+  //   region: item.region,
+  //   summonerName: item.summonerName || item.opgg || "what?",
+  //   role: item.role,
+  //   champions: item.champions || null,
+  //   timezone: item.timezone,
+  //   info: item.info || null,
+  //   createdAt: item.createdAt.toISOString(),
+  //   discordName: item.discordName,
+  //   discordId: item.discordId,
+  //   accepted: item.accepted?.toISOString() || null,
+  //   completed: item.completed?.toISOString() || null,
+  //   mentor: item.mentor?.discordName || null,
+  //   remarks: item.remarks || null,
+  // }));
 }
